Implement autogrow for author WYSIWYG
WYSIWYG Editor auto increases height as content is either A) loaded or B) added. This makes it behave much more like Google Docs, a system our teachers are familiar with.

Note, this requires the `autogrow` package for CKEditor which I think is a default package already installed on the server but we'd need to double check before deploying.

I've found one unintended consequence in that this causes the default height of ckeditor instances to be slightly larger than before, but we actually had several people ask for a default height increase anyway so this works out!