Change "Training" to "Teacher Support"
This is a much bigger change. For posterity, it makes more sense to change this throughout the site meaning it requires renaming some view controllers and making sure the correct view is loaded. I've renamed everything using the usual case-structure from training -> teacherSupport.

This also adds a block with an identical id on the Home page that links to the new teacher support page. This shouldn't be a problem since it doesn't affect layout or reference or anything.