Add poetry support
Poetry as a dependency and `venv` manager is an incredibly easy to way to start development. Particularly, if this project is ever passed on.

https://python-poetry.org

If one were to install poetry...one need only ensure a compatible Python version (in this case, 3.7.0) is installed and selected (could use `pyenv`) and then simply run `poetry install`. This installs dependencies in a virtual environment stored locally. You can then run `poetry shell` to enter the virtual environment. This would be an alternative to the current local-dev support...BUT, i believe the two can live together if need be.

It also allows for exporting requirements.txt (see previous commit).