Add django-tinymce==2.80
Note that this is a legacy requirement that is NO LONGER NEEDED.

I'm adding it back in since in the repo, we have maintained ALL of our migration files, two of which depend on importing TinyMCE (the only other file is `forms.py`).

We should probably set a standard for how to squash migrations.